requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
youtube-transcript-api==1.2.3
requests==2.31.0
python-dotenv==1.0.0
openai>=1.12.0,<2.0.0
httpx>=0.25.0
googlemaps==4.10.0
